+++ src/client/java/io/github/tr100000/researcher/mixin/client/compat/jei/CraftingRecipeCategoryMixin.java
@@ -26,11 +26,11 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(CraftingRecipeCategory.class)
 public class CraftingRecipeCategoryMixin {
-    @Shadow @Final
-    public static int height;
-
     @Unique
     private static final Identifier ERROR_TEXTURE = ModUtils.id("textures/gui/error_overlay.png");
+
+    @Shadow @Final
+    public static int height;
 
     @Inject(method = "draw(Lnet/minecraft/world/item/crafting/RecipeHolder;Lmezz/jei/api/gui/ingredient/IRecipeSlotsView;Lnet/minecraft/client/gui/GuiGraphicsExtractor;DD)V", at = @At("TAIL"))
     private void draw(RecipeHolder<CraftingRecipe> recipeHolder, IRecipeSlotsView recipeSlotsView, GuiGraphicsExtractor guiGraphics, double mouseX, double mouseY, CallbackInfo ci) {