+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+### Added
+- Added "Only When Holding Shift" as an option for enabling resize handles in the research screen
+- Locked recipes are now indicated in RRV properly
+
 ### Changed
 - Slightly improved performance of graph layout algorithm
+- Increased the height of the search bar in the research list
+
+### Fixed
+- Resize-snapping the research sidebar now rounds to the nearest acceptable width instead of always rounding down
+- The resize "bars" now stay highlighted while dragging, even if the mouse is far away somehow